+++ src/app.js
@@ -14,7 +14,7 @@ const gitRouter = require('./routes/gitauth');
 const emailVerificationRouter = require('./routes/EmailVerification');
 const resetPasswordRouter = require('./routes/resetPassword');
 const { connectDB } = require('./controllers/db');
-const { errorHandler, unknownRoutes } = require('./utils/middleware');
+const { errorHandler, unknownRoutes } = require('./middlewares/middleware');
 const { authorizeUser } = require('./controllers/auth');
 const swaggerUi = require('swagger-ui-express');
 const passport = require('passport');
