+++ src/filestructure
@@ -7,4 +7,5 @@ src__
     |__ services - all core logic should be written here
     |__ controllers - all responses should be handled here using the right services
     |__ routes - all specific route definitions here
+    |__ tests - all test code files resides here
 ...
